+++ runelite-client/src/main/java/net/runelite/client/plugins/bankvalue/BankCalculation.java
@@ -115,7 +115,7 @@ class BankCalculation
 
 			if (config.showHA())
 			{
-				int alchValue = itemManager.getAlchValue(widgetItem.getId());
+				long alchValue = itemManager.getAlchValue(widgetItem.getId());
 
 				if (alchValue > 0)
 				{
