+++ runelite-client/src/main/java/net/runelite/client/plugins/examine/ExaminePlugin.java
@@ -289,7 +289,7 @@ public class ExaminePlugin extends Plugin
 		quantity = Math.max(1, quantity);
 		int itemCompositionPrice = itemComposition.getPrice();
 		final int gePrice = itemManager.getItemPrice(id);
-		final int alchPrice = itemCompositionPrice <= 0 ? 0 : itemManager.getAlchValue(id);
+		final long alchPrice = itemCompositionPrice <= 0 ? 0 : itemManager.getAlchValue(id);
 
 		if (gePrice > 0 || alchPrice > 0)
 		{
